fix : tripresponse 수정

See merge request s12-webmobile4-sub1/S12P11D210!29

+++ BackEnd/TravelTogether/src/main/java/com/travel/together/TravelTogether/trip/repository/TripRepository.java
@@ -10,9 +10,9 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface TripRepository extends JpaRepository<Trip, Long> {
+public interface TripRepository extends JpaRepository<Trip, Integer> {
     @Query("SELECT t FROM Trip t JOIN FETCH t.creator WHERE t.id = :tripId")
-    Optional<Trip> findByIdWithCreator(@Param("tripId") Long tripId);
+    Optional<Trip> findByIdWithCreator(@Param("tripId") Integer tripId);
 
-    List<Trip> findByCreatorIdOrderByCreatedAtDesc(Long creatorId);
+    List<Trip> findByCreatorIdOrderByCreatedAtDesc(Integer creatorId);
 }
